requirements.txt
torch>=1.9.0
torchvision>=0.10.0
clip @ git+https://github.com/openai/CLIP.git
numpy>=1.21.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0
Pillow>=8.0.0
opencv-python>=4.5.0
scikit-image>=0.18.0 